Add sidekiq-cron gem and SettleReceivablesWorker

+++ lib/types.rb
@@ -5,7 +5,7 @@ module Types
   Gateway = Types::Coercible::String.enum(*::Gateways::LIST.map(&:to_s))
 
   module PaymentTransactions
-    Amount = Types::Coercible::Decimal.constrained(gt: 0)
+    Amount = Types::Coercible::Float.constrained(gt: 0)
     Installment = Types::Coercible::Integer.constrained(gteq: 1, lteq: 12)
     PaymentMethod = Types::Coercible::String.enum(*::PaymentTransaction::PAYMENT_METHODS)
     Status = Types::Coercible::String.enum(*::PaymentTransaction::STATUSES)
